requirements.txt
# Core dependencies
streamlit==1.35.0
langchain==0.2.3
langchain-google-genai==1.0.2
python-dotenv==1.1.0

# Embeddings + Vector Store
faiss-cpu==1.7.4
sentence-transformers==2.6.1
scikit-learn==1.1.3

# File loaders
PyMuPDF==1.24.3
docx2txt==0.8

# Evaluation
evaluate==0.4.1
bert-score==0.3.13

# LLM + Tokenizer
transformers==4.41.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
